requirements.txt
spacy
pandas
numpy
matplotlib
seaborn
nltk
scikit-learn
pytorch-crf
transformers